fixed changes
