+++ src/main/java/com/example/sb_assign_16_05_23/entity/Subject.java
@@ -29,3 +29,4 @@ public class Subject {
 
 // constructors, getters and setters
 }
+
