+++ src/main/java/com/example/sb_assign_16_05_23/service/impl/StudentServiceImpl.java
@@ -4,6 +4,7 @@ import com.example.sb_assign_16_05_23.dto.ResponseDTO;
 import com.example.sb_assign_16_05_23.entity.Student;
 import com.example.sb_assign_16_05_23.repository.StudentRepository;
 import com.example.sb_assign_16_05_23.service.StudentService;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
@@ -16,15 +17,16 @@ public class StudentServiceImpl implements StudentService{
     @Autowired
     StudentRepository studentRepository;
 
-    @Qualifier("mapperServiceImpl")
     @Autowired
-    MapperServiceImpl mapperService;
+    ResponseDTO<?> responseDTO;
 
     @Override
     public ResponseDTO<?> getAllStudents() {
         List<Student> students = studentRepository.findAll();
 
-        return mapperService.getResponseDTO(students, "All Students Retrieved From Database");
+        System.out.println(responseDTO.getResponseDTO(students, "All Students Retrieved From Database").toString());
+
+        return responseDTO.getResponseDTO(students, "All Students Retrieved From Database");
     }
 }
