+++ src/main/java/com/example/sb_assign_16_05_23/service/impl/TeacherServiceImpl.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -23,20 +24,19 @@ TeacherServiceImpl implements TeacherService {
     TeacherRepository teacherRepository;
 
     @Autowired
-    List<Teacher> teacherList;
+    ModelMapper modelMapper;
 
-    @Qualifier("mapperServiceImpl")
     @Autowired
-    MapperServiceImpl mapperService;
+    ResponseDTO<?> responseDTO;
 
 
 
     public Teacher teacherDTOToTeacher(TeacherDTO teacherData){
 
-        Teacher teacher =  (Teacher) mapperService.Mapper(teacherData, Teacher.class);
+        Teacher teacher =  modelMapper.map(teacherData, Teacher.class);
         List<Subject> subjects = teacherData.getSubjects()
                 .stream()
-                .map(subject -> (Subject)mapperService.Mapper(subject, Subject.class))
+                .map(subject -> modelMapper.map(subject, Subject.class))
                 .collect(Collectors.toList());
 
         subjects.forEach(subject -> subject.setTeacher(teacher));
@@ -51,25 +51,27 @@ TeacherServiceImpl implements TeacherService {
 
         List<Teacher> teachers = teacherRepository.findAll();
 
-        return mapperService.getResponseDTO(teachers, "All Teachers Retrieved From Database");
+        return responseDTO.getResponseDTO(teachers, "All Teachers Retrieved From Database");
     }
 
     @Override
     public ResponseDTO<?> setTeachers(TeacherDTO teacherData){
 
         Teacher teacher = teacherDTOToTeacher(teacherData);
-        List<TeacherDTO> teacherDTO = Collections.singletonList((TeacherDTO) mapperService.Mapper(teacher, TeacherDTO.class));
+        List<TeacherDTO> teacherDTO = Collections.singletonList(modelMapper.map(teacher, TeacherDTO.class));
 
-        return mapperService.getResponseDTO(teacherDTO, "Teacher added to Database");
+        return responseDTO.getResponseDTO(teacherDTO, "Teacher added to Database");
     }
 
     @Override
     public ResponseDTO<?> setAll(List<TeacherDTO> teacherData) {
 
+        List<Teacher> teacherList = new ArrayList<>();
+
         for(TeacherDTO teacherD: teacherData){
             teacherList.add(teacherDTOToTeacher(teacherD));
         }
-        return mapperService.getResponseDTO(teacherList, "List of Teachers added to Database");
+        return responseDTO.getResponseDTO(teacherList, "List of Teachers added to Database");
 
     }
 }
